+++ Funcky/Sequence/Return.cs
@@ -5,6 +5,7 @@ namespace Funcky
 {
     public static partial class Sequence
     {
-        public static IEnumerable<TItem> Return<TItem>(TItem item) => Enumerable.Repeat(item, 1);
+        public static IEnumerable<TItem> Return<TItem>(TItem item)
+            => Enumerable.Repeat(item, 1);
     }
 }